onAuthStateChanged optimised
