+++ src/Components/AppLayout.js
@@ -2,11 +2,6 @@ import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import Browse from "./Browse";
 import Login from "./Login";
 import HomePage from "./HomePage";
-import { useEffect } from "react";
-import { onAuthStateChanged } from "firebase/auth";
-import { auth } from "../Utils/firebase";
-import { useDispatch } from "react-redux";
-import { addUser, removeUser } from "../Utils/userSlice";
 
 const appRouter = createBrowserRouter([
   {
@@ -28,26 +23,8 @@ const appRouter = createBrowserRouter([
 ]);
 
 const AppLayout = () => {
-  const dispatch = useDispatch();
   return (
     <div>
-      {useEffect(() => {
-        onAuthStateChanged(auth, (user) => {
-          if (user) {
-            const { uid, email, displayName, photoURL } = user;
-            dispatch(
-              addUser({
-                uid: uid,
-                email: email,
-                displayName: displayName,
-                photoURL: photoURL,
-              })
-            );
-          } else {
-            dispatch(removeUser());
-          }
-        });
-      }, [])}
       <RouterProvider router={appRouter} />
     </div>
   );
